Fixed bug causing LJ music to be title of summary (sort of)
So LJ does this annoying thing where the 'title' field in its RSS feeds
is actually the input you gave for 'music'.  But, so is 'lj_music'.  I
accidentally discovered that the atom feed doesn't have this problematic
behavior, so I switched the default setting over to it.  There was much
rejoicing.